+++ projects/np-ui-lib/src/lib/np-number-box/np-number-box.component.ts
@@ -33,7 +33,7 @@ export class NpNumberBoxComponent implements ControlValueAccessor, Validator {
   @Input() placeholder = '';
   @Input() readOnly: boolean;
   @Input() autoFocus: boolean;
-  @Input() tabIndex: number;
+  @Input() tabIndex = 0;
   @Input() styleClass: string;
   @Input() inputId = `np-number-box_${NpNumberBoxComponent.controlCount++}`;
 
@@ -58,7 +58,6 @@ export class NpNumberBoxComponent implements ControlValueAccessor, Validator {
     if (v !== this.innerValue) {
       this.innerValue = v;
       this.onChangeCallback(v);
-      this.onTouchedCallback();
       this.onChange.emit(v);
     }
   }