+++ projects/np-ui-lib/src/lib/np-switch/np-switch.component.ts
@@ -22,7 +22,7 @@ export class NpSwitchComponent implements ControlValueAccessor {
   @Input() falseLabelText: string;
   @Input() readOnly: boolean;
   @Input() autoFocus: boolean;
-  @Input() tabIndex: number;
+  @Input() tabIndex = 0;
   @Input() styleClass: string;
   @Input() inputId = `np-switch_${NpSwitchComponent.controlCount++}`;
 
@@ -46,7 +46,6 @@ export class NpSwitchComponent implements ControlValueAccessor {
     if (v !== this.innerValue) {
       this.innerValue = v;
       this.onChangeCallback(v);
-      this.onTouchedCallback();
       this.onChange.emit(v);
     }
   }