+++ projects/np-ui-lib/src/lib/np-time-picker/np-time-picker.component.ts
@@ -29,7 +29,7 @@ export class NpTimePickerComponent implements ControlValueAccessor, AfterViewIni
   @Input() placeholder = '';
   @Input() readOnly: boolean;
   @Input() autoFocus: boolean;
-  @Input() tabIndex: number;
+  @Input() tabIndex = 0;
   @Input() styleClass: string;
   @Input() inputId = `np-time-picker_${NpTimePickerComponent.controlCount++}`;
 
@@ -92,7 +92,6 @@ export class NpTimePickerComponent implements ControlValueAccessor, AfterViewIni
     if (v !== this.innerValue) {
       this.innerValue = v;
       this.onChangeCallback(v);
-      this.onTouchedCallback();
       this.onChange.emit(v);
     }
   }